refactor: move user queries to correct directory
