+++ projects/client/src/lib/requests/queries/users/currentUserSettingsQuery.spec.ts
