+++ projects/client/src/lib/requests/queries/users/currentUserSettingsQuery.ts
@@ -7,7 +7,7 @@ import type {
 import { DEFAULT_COVER } from '$lib/utils/constants.ts';
 import { findDefined } from '$lib/utils/string/findDefined.ts';
 import { prependHttps } from '$lib/utils/url/prependHttps.ts';
-import { api, type ApiParams } from '../../../requests/_internal/api.ts';
+import { api, type ApiParams } from '../../_internal/api.ts';
 
 export type UserSettings = {
   id: string;
