+++ projects/client/src/lib/requests/queries/users/currentUserWatchlistQuery.ts
@@ -3,7 +3,7 @@ import type {
   WatchlistedShowsResponse,
 } from '$lib/api.ts';
 import { InvalidateAction } from '$lib/requests/models/InvalidateAction.ts';
-import { api, type ApiParams } from '../../../requests/_internal/api.ts';
+import { api, type ApiParams } from '../../_internal/api.ts';
 
 type WatchlistedMedia = {
   id: number;
