+++ projects/client/src/routes/+layout.ts
@@ -1,7 +1,7 @@
 import type { LayoutLoad } from './$types';
 
 import { browser } from '$app/environment';
-import { currentUserSettingsQuery } from '$lib/features/auth/queries/currentUserSettingsQuery.ts';
+import { currentUserSettingsQuery } from '$lib/requests/queries/users/currentUserSettingsQuery.ts';
 import { QueryClient } from '@tanstack/svelte-query';
 
 export const load: LayoutLoad = async ({ data, fetch }) => {
